+++ src/Pim/Bundle/BaseConnectorBundle/Writer/Doctrine/ProductWriter.php
@@ -6,8 +6,7 @@ use Akeneo\Bundle\BatchBundle\Entity\StepExecution;
 use Akeneo\Bundle\BatchBundle\Item\AbstractConfigurableStepElement;
 use Akeneo\Bundle\BatchBundle\Item\ItemWriterInterface;
 use Akeneo\Bundle\BatchBundle\Step\StepExecutionAwareInterface;
-use Akeneo\Component\StorageUtils\Saver\SaverInterface;
-use Pim\Bundle\CatalogBundle\Manager\MediaManager;
+use Akeneo\Component\StorageUtils\Saver\BulkSaverInterface;
 use Pim\Bundle\CatalogBundle\Model\ProductInterface;
 use Pim\Bundle\TransformBundle\Cache\CacheClearer;
 use Pim\Bundle\VersioningBundle\Manager\VersionManager;
@@ -23,9 +22,6 @@ class ProductWriter extends AbstractConfigurableStepElement implements
     ItemWriterInterface,
     StepExecutionAwareInterface
 {
-    /**  @var MediaManager */
-    protected $mediaManager;
-
     /** @var VersionManager */
     protected $versionManager;
 
@@ -38,24 +34,21 @@ class ProductWriter extends AbstractConfigurableStepElement implements
     /** @var bool */
     protected $realTimeVersioning = true;
 
-    /** @var SaverInterface */
+    /** @var BulkSaverInterface */
     protected $productSaver;
 
     /**
      * Constructor
      *
-     * @param MediaManager   $mediaManager
-     * @param CacheClearer   $cacheClearer
-     * @param VersionManager $versionManager
-     * @param SaverInterface $productSaver
+     * @param CacheClearer       $cacheClearer
+     * @param VersionManager     $versionManager
+     * @param BulkSaverInterface $productSaver
      */
     public function __construct(
-        MediaManager $mediaManager,
         CacheClearer $cacheClearer,
         VersionManager $versionManager,
-        SaverInterface $productSaver
+        BulkSaverInterface $productSaver
     ) {
-        $this->mediaManager   = $mediaManager;
         $this->cacheClearer   = $cacheClearer;
         $this->versionManager = $versionManager;
         $this->productSaver   = $productSaver;
@@ -106,7 +99,6 @@ class ProductWriter extends AbstractConfigurableStepElement implements
         foreach ($items as $item) {
             $this->incrementCount($item);
         }
-        $this->mediaManager->handleAllProductsMedias($items);
         $this->productSaver->saveAll($items, ['recalculate' => false]);
 
         $this->cacheClearer->clear();