+++ src/Pim/Bundle/TransformBundle/Transformer/Property/MediaTransformer.php
@@ -2,6 +2,8 @@
 
 namespace Pim\Bundle\TransformBundle\Transformer\Property;
 
+use Akeneo\Component\FileStorage\Exception\FileTransferException;
+use Akeneo\Component\FileStorage\RawFile\RawFileStorerInterface;
 use Pim\Bundle\TransformBundle\Exception\PropertyTransformerException;
 use Pim\Bundle\TransformBundle\Transformer\ColumnInfo\ColumnInfoInterface;
 use Symfony\Component\HttpFoundation\File\Exception\FileNotFoundException;
@@ -16,15 +18,15 @@ use Symfony\Component\HttpFoundation\File\File;
  */
 class MediaTransformer implements PropertyTransformerInterface, EntityUpdaterInterface
 {
-    /** @var string */
-    protected $mediaClass;
+    /** @var RawFileStorerInterface */
+    protected $storer;
 
     /**
-     * @param string $mediaClass
+     * @param RawFileStorerInterface $storer
      */
-    public function __construct($mediaClass)
+    public function __construct(RawFileStorerInterface $storer)
     {
-        $this->mediaClass = $mediaClass;
+        $this->storer = $storer;
     }
 
     /**
@@ -39,9 +41,17 @@ class MediaTransformer implements PropertyTransformerInterface, EntityUpdaterInt
         }
 
         try {
-            $file = new File($value);
+            $rawFile = new File($value);
+            $file = $this->storer->store($rawFile, 'storage');
         } catch (FileNotFoundException $e) {
-            throw new PropertyTransformerException('File not found: "%value%"', array('%value%' => $value));
+            throw new PropertyTransformerException('File not found: "%value%"', ['%value%' => $value]);
+        } catch (FileTransferException $e) {
+            throw new PropertyTransformerException('Impossible to transfer the file "%value%"', ['%value%' => $value]);
+        } catch (\Exception $e) {
+            throw new PropertyTransformerException(
+                'An error occurred during the process of the file "%value%"',
+                ['%value%' => $value]
+            );
         }
 
         return $file;
@@ -56,11 +66,6 @@ class MediaTransformer implements PropertyTransformerInterface, EntityUpdaterInt
             return;
         }
 
-        $media = $object->getMedia();
-        if (!$media) {
-            $media = new $this->mediaClass();
-            $object->setMedia($media);
-        }
-        $media->setFile($data);
+        $object->setMedia($data);
     }
 }